refactor: altera retorno do endpoint de deletar usuário

+++ src/main/java/com/example/todo/services/UserService.java
@@ -58,7 +58,9 @@ public class UserService {
   }
 
   public void deleteUser(Long id) {
-    getOneUser(id);
+    if (!repository.existsById(id)) {
+      throw new NotFoundException("User not found");
+    }
     repository.deleteById(id);
   }
 }
